feat: change last stack fun

+++ create_stack.c
@@ -29,7 +29,8 @@ bool	create_stack(t_stack_node **a, char **argv)
 		}
 		// printf("last stack node: %d\n ", (int)ft_stacklast(a)->value);
 		// a is big node(a stack, how can I have access  ? );
-		printf("value A : %d\n", (int)ft_stacklast(a)->value);
+		// below code has segmentation error
+		printf("value A : %d\n", ft_stacklast(a)->value);
 		free(node);
 		i++;
 	}
